# openapi/ortho-v1.yaml
openapi: 3.1.0
info:
  title: ORTHO-32 API
  version: 1.0.0
  description: |
    Canonical versioned REST/SSE contract all ORTHO clients talk to.
    All responses are wrapped in Envelope { id, type, timestamp, source, correlation_id, data }.
    cycles fields are integer architectural cycles, NEVER wall-clock ms or float.
    VerificationStatus has exactly 7 disjoint values, never interchangeable.
    Elevated scopes: hardware.reset, marketplace.install require explicit grant.
servers:
  - url: http://localhost:7032
    description: Local development (HTTP transport)
  - url: https://ortho.local:7032
    description: Remote HTTP/2 transport
  - url: ipc://ortho32-api.sock
    description: Local IPC transport (desktop co-located)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Scoped capability token from POST /api/v1/auth/session
  schemas:
...
    Cycles:
      type: integer
      minimum: 0
      description: "Architectural cycle count - integer. NEVER float, NEVER wall-clock ms."
      example: 1024
...
paths:
...
  /api/v1/traces/capture:
    post:
      summary: Capture trace
      operationId: captureTrace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                durationCycles: { $ref: "#/components/schemas/Cycles" }
      responses: { '201': { description: captured } }
  /api/v1/traces/{id}:
    get:
      summary: Get trace
      operationId: getTrace
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/traces/{id}/replay:
    post:
      summary: Replay trace
      operationId: replayTrace
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: replayed } }
  /api/v1/traces/compare:
    post:
      summary: Compare traces
      operationId: compareTraces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [a, b]
              properties:
                a: { type: string }
                b: { type: string }
      responses: { '200': { description: compared } }
...